Move condition-related functions to a separate file
